- Move statistics to its own view
