+++ app/Helpers/GeneralHelper.php
@@ -130,54 +130,4 @@ class GeneralHelper
 
         return $result;
     }
-
-    /**
-     * Returns Ban statistics
-     *
-     * @return array
-     */
-    public static function getBanStats(): array
-    {
-        $key = 'ban_statistics';
-        if (Cache::has($key)) {
-            return Cache::get($key, []);
-        }
-
-        $stats = Ban::query()->fromSub(function ($query) {
-            $query->from('user_bans')->select([
-                DB::raw('FROM_UNIXTIME(`timestamp`, \'%Y-%m-%d\') AS `date`'),
-            ])->orderByDesc('timestamp')->groupBy('ban_hash');
-        }, 'bans')->select([
-            DB::raw('COUNT(`date`) as `count`'),
-            'date',
-        ])->groupBy('date')->get()->toArray();
-
-        $map = [];
-        foreach ($stats as $row) {
-            $map[$row['date']] = $row['count'];
-        }
-
-        $complete = [];
-        for ($x = 30; $x >= 0; $x--) {
-            $t = date('Y-m-d', strtotime('-' . $x . ' day'));
-            if (!isset($map[$t])) {
-                $complete[$t] = 0;
-            } else {
-                $complete[$t] = $map[$t];
-            }
-        }
-
-        $data = [
-            'labels' => [],
-            'data'   => [],
-        ];
-        foreach ($complete as $label => $value) {
-            $data['labels'][] = $label;
-            $data['data'][] = $value;
-        }
-
-        Cache::put($key, $data, 6 * 60 * 60);
-
-        return $data;
-    }
 }
